+++ lib/swagger_docs_generator/parser/parser.rb
@@ -41,3 +41,4 @@ end
 require 'swagger_docs_generator/parser/controller'
 require 'swagger_docs_generator/parser/action'
 require 'swagger_docs_generator/parser/definition'
+require 'swagger_docs_generator/parser/readme'